\name{col_format_dif}
\alias{col_format_dif}
\title{Column Format for Differences}
\description{
This function creates a column format for differences.
}
\usage{
col_format_dif(digits = 1)
}
\arguments{
\item{digits}{Number of digits to display.}
}
\value{
A column format object for differences.
}
\examples{
col_format_dif(2)
}
