\name{col_format_dollar}
\alias{col_format_dollar}
\title{Column Format for Dollar Values}
\description{
This function creates a column format for dollar values.
}
\usage{
col_format_dollar(digits)
}
\arguments{
\item{digits}{Number of digits to display.}
}
\value{
A column format object for dollar values.
}
\examples{
col_format_dollar(2)
}
