\name{col_format_per}
\alias{col_format_per}
\title{Column Format for Percentage Values}
\description{
This function creates a column format for percentage values.
}
\usage{
col_format_per(digits = 1)
}
\arguments{
\item{digits}{Number of digits to display.}
}
\value{
A column format object for percentage values.
}
\examples{
col_format_per(2)
}
