% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_theme.R
\name{customReactable}
\alias{customReactable}
\title{Create Custom Reactable}
\usage{
customReactable(
  data,
  ns,
  columns,
  paginateSubRows = TRUE,
  resizable = TRUE,
  striped = FALSE,
  wrap = FALSE,
  pagination = TRUE,
  showPagination = TRUE,
  showPageSizeOptions = TRUE,
  filterable = TRUE,
  highlight = FALSE,
  height = "500",
  width = "100\%",
  selection = "multiple",
  ...
)
}
\arguments{
\item{data}{A data frame to display in the reactable.}

\item{ns}{A namespace function to prefix input IDs for Shiny modules.}

\item{columns}{A list of column definitions for the reactable.}

\item{paginateSubRows}{Logical, whether to paginate sub-rows. Default is TRUE.}

\item{resizable}{Logical, whether columns are resizable. Default is TRUE.}

\item{striped}{Logical, whether to display striped rows. Default is FALSE.}

\item{wrap}{Logical, whether to wrap text in cells. Default is FALSE.}

\item{pagination}{Logical, whether to enable pagination. Default is TRUE.}

\item{showPagination}{Logical, whether to show pagination controls. Default is TRUE.}

\item{showPageSizeOptions}{Logical, whether to show page size options. Default is TRUE.}

\item{filterable}{Logical, whether to enable filtering. Default is TRUE.}

\item{highlight}{Logical, whether to highlight rows on hover. Default is FALSE.}

\item{height}{The height of the table in pixels or as a percentage. Default is "500".}

\item{width}{The width of the table in pixels or as a percentage. Default is "100%".}

\item{selection}{The selection mode for rows ("multiple", "single", or "none"). Default is "multiple".}

\item{...}{Additional arguments passed to the reactable function.}
}
\description{
This function creates a custom reactable with specified common styles for light theme apps.
}
