% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable_table_ui.R
\name{editable_table_ui}
\alias{editable_table_ui}
\title{UI function for the editable table Shiny module}
\usage{
editable_table_ui(id)
}
\arguments{
\item{id}{A string. The module namespace ID.}
}
\value{
A named list of \code{shiny.tag} objects.
}
\description{
UI function for the editable table Shiny module
}
