% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_theme.R
\docType{data}
\name{fontsize}
\alias{fontsize}
\title{Font Size}
\format{
An object of class \code{character} of length 1.
}
\usage{
fontsize
}
\description{
This variable defines the font size for the reactable theme.
}
\keyword{datasets}
