\name{format_values_1k}
\alias{format_values_1k}
\title{Format Values in Thousands}
\description{
This function multiplies a given value by 1000 and formats it.
}
\usage{
format_values_1k(value)
}
\arguments{
\item{value}{A numeric value to be formatted.}
}
\value{
A formatted value.
}
\examples{
format_values_1k(10)
}
