% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade-utils.R
\name{gradeIndex}
\alias{gradeIndex}
\title{Grade ordering index}
\usage{
gradeIndex(key)
}
\arguments{
\item{key}{Character. One or more grade keys.}
}
\value{
Integer vector.
}
\description{
Converts a grade key to a numeric index for comparison.
PK=0, K=1, 01=2, 02=3, ..., 12=13.
}
\examples{
gradeIndex("PK")  # 0
gradeIndex("K")   # 1
gradeIndex("05")  # 6
gradeIndex("12")  # 13

}
