% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_theme.R
\name{muiDependency}
\alias{muiDependency}
\title{Dependencies for Material-UI Components}
\usage{
muiDependency()
}
\description{
This function returns the necessary HTML dependencies to use Material-UI components in a Shiny application.
}
