% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactablePlus-package.R
\docType{package}
\name{reactablePlus-package}
\alias{reactablePlus}
\alias{reactablePlus-package}
\title{reactablePlus: Enhanced Interactive Tables with Inline Editors for Shiny}
\description{
Extends 'reactable' with inline editing widgets (pickers, popovers, text inputs) and a declarative configuration layer for building editable tables in Shiny applications. Provides a generic editable table module driven by row and column specifications, plus a library of cell-level input widgets with custom Shiny bindings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewcurrier/reactablePlus}
  \item \url{https://matthewcurrier.github.io/reactablePlus/}
  \item Report bugs at \url{https://github.com/matthewcurrier/reactablePlus/issues}
}

}
\author{
\strong{Maintainer}: Matthew Currier \email{currier.matthew@gmail.com}

Authors:
\itemize{
  \item Matthew Currier \email{currier.matthew@gmail.com}
}

}
\keyword{internal}
