% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_theme.R
\name{style.red.green}
\alias{style.red.green}
\title{Style for Positive and Negative Values}
\usage{
style.red.green(value)
}
\arguments{
\item{value}{A numeric value to be styled.}
}
\description{
This function returns a style list for positive and negative values.
}
