% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_theme.R
\name{theme}
\alias{theme}
\title{Create Reactable Theme}
\usage{
theme(paddingBottom = "0.5rem", paddingTop = "0.5rem")
}
\arguments{
\item{paddingBottom}{Bottom padding for the search input style.}

\item{paddingTop}{Top padding for the search input style.}
}
\description{
This function creates a custom reactable theme with specified styles for various elements.
}
