% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear-popover.R
\name{updateGearPopoverInput}
\alias{updateGearPopoverInput}
\title{Update a gear popover from the server}
\usage{
updateGearPopoverInput(
  session = getDefaultReactiveDomain(),
  inputId,
  value = NULL
)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{Character. The input ID to update.}

\item{value}{Named list of toggle values to update.}
}
\value{
Called for its side effect; returns invisibly.
}
\description{
Update a gear popover from the server
}
