% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeschool-picker.R
\name{updateHomeschoolPickerInput}
\alias{updateHomeschoolPickerInput}
\title{Update a homeschool picker from the server}
\usage{
updateHomeschoolPickerInput(
  session = getDefaultReactiveDomain(),
  inputId,
  value = NULL
)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{Character. The input ID to update.}

\item{value}{New value — a list with details or \code{NULL} to turn off.}
}
\value{
Called for its side effect; returns invisibly.
}
\description{
Update a homeschool picker from the server
}
